api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
huggingface_hub>=0.22.0
python-multipart>=0.0.9
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
